src/main/java/com/qwest/backend/business/impl/UserDetailsServiceImpl.java
package com.qwest.backend.business.impl;

import com.qwest.backend.domain.Author;
import com.qwest.backend.repository.AuthorRepository;
import org.springframework.security.core.authority.SimpleGrantedAuthority;
import org.springframework.security.core.userdetails.User;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;
import java.util.Collections;
import java.util.List;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {

    private final AuthorRepository authorRepository;

    public UserDetailsServiceImpl(AuthorRepository authorRepository) {
        this.authorRepository = authorRepository;
    }

    @Override
    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
        Author author = authorRepository.findByEmail(email)
                .orElseThrow(() -> new UsernameNotFoundException("No user found with email: " + email));

        if (author.getRole() == null) {
            throw new IllegalArgumentException("User has no roles assigned");
        }

        List<SimpleGrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority("ROLE_" + author.getRole().name()));

        return new User(author.getEmail(), author.getPasswordHash(), authorities);
    }

}
